Match func_800FB97C_2A07AC from ovl_61

+++ src/engine/pad.c
@@ -225,7 +225,7 @@ void func_800141C4(void) {
     D_800D12CC = 0;
 }
 
-u8 func_800141FC(s16 arg0) {
+u32 func_800141FC(s16 arg0) {
     return D_800D12BA[arg0];
 }
